Struct dexter::Experiment

#[must_use]
pub struct Experiment<'a, P, Cr: Clone, Nr: Clone> {
    // some fields omitted
}

A struct for building Dexter experiments.

Methods

impl<'a, P, Cr: Clone, Nr: Clone> Experiment<'a, P, Cr, Nr>

fn new<C, N>(name: &'static str, current_subject: C, new_subject: N) -> Self where C: FnMut(&P) -> Cr + 'a, N: FnMut(&P) -> Nr + 'a

Constructs a new Experiment with a current and given subject.

fn setup<S>(self, setup: S) -> Self where S: FnMut(P) -> P + 'a

Adds a setup step to the experiment.

The setup function can alter the parameter that's passed into the experiment before it is passed to the run_if closure, the ignore_if closure, and the two subject closures.

fn run_if<R>(self, run_if: R) -> Self where R: FnMut(&P) -> bool + 'a

Adds a check step that will disable the experiment in certain cases.

If the passed closure returns false when passed the experiment's parameter, then the experiment will return the current subject's result without publishing.

fn ignore_if<I>(self, ignore_if: I) -> Self where I: FnMut(&P) -> bool + 'a

Adds an check step that will ignore mismatches in certain cases.

If the passed closure returns true when passed the experiment's parameter, then the result will have a MatchType of Ignored.

fn carry_out<Pub: Publisher<P, Cr, Nr>>(self, param: P, publisher: &mut Pub) -> Cr

Carry out the experiment given a parameter and a publisher.

Returns the result of the current subject closure.