Trait dexter::Publisher

pub trait Publisher<P, Cr: Clone, Nr: Clone> {
    fn publish(&mut self, result: ExperimentResult<Cr, Nr>);
    fn compare(&mut self, current_result: &Cr, new_result: &Nr) -> bool;

    fn enabled(&mut self) -> bool { ... }
}

Trait for publishers, which are used by Dexter to store results of experiments.

Required Methods

fn publish(&mut self, result: ExperimentResult<Cr, Nr>)

Publish the result of an experiment.

fn compare(&mut self, current_result: &Cr, new_result: &Nr) -> bool

Comparison function for the results of the subjects.

This function should return true if the results "match," and false if they do not.

Provided Methods

fn enabled(&mut self) -> bool

Only run the experiment in some cases.

If enabled returns false, then the result of the current subject is used and no results are published. This is meant to be an inexpensive function that gets called when every experiment runs.